getVideoAdapters
